+++ NEWS.md
@@ -3,6 +3,7 @@
  * "Copy Link" for bookmarks
  * "Add bookmark": use current url if clipboard doesn't contains url
  * Handle feed title. Put into window title.
+ * New settings option: network timeout
 
 # Version 0.0.5
  * Context menu for download-manager